% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage20x-package.R
\docType{package}
\name{testpackage20x-package}
\alias{testpackage20x}
\alias{testpackage20x-package}
\title{testpackage20x: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
